Update simulator.rst
Question! In some of the documentation, you put the name of buttons in quotes. EX: This is an example of the "Push" button. And sometimes there are no quotes at all. Do you have a preference? Generally, when I document buttons I make them **bold**. What do you think?

Everything else in this article is good to go!